requirements.txt
# Updated requirements.txt with yfinance as primary data source

# Core data analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# PRIMARY DATA SOURCE - yfinance for real stock prices (NO API KEY NEEDED!)
yfinance>=0.2.18

# API and web requests (kept for backwards compatibility)
requests>=2.31.0
python-dotenv>=0.19.0

# Visualization - Core
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Scientific computing
scikit-learn>=1.3.0

# Development and testing
jupyter>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional for advanced features
pillow>=9.0.0

# Configuration and utilities
PyYAML>=6.0
psutil>=5.9.0
tenacity>=8.1.0
colorlog>=6.7.0

# Dashboard Framework (Streamlit)
streamlit>=1.28.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6

# Alternative Financial Data APIs (optional backups)
alpha-vantage>=2.3.1  # Keep for backwards compatibility
# Note: yfinance is free and doesn't require API keys!

# Async Processing
aiohttp>=3.8.0
websocket-client>=1.6.1

# Code Quality (Development)
black>=23.0.0
flake8>=6.0.0

# Professional Dashboard Features
streamlit-echarts>=0.4.0
plotly-express>=0.4.1

# For real-time updates
schedule>=1.2.0

# For enhanced error handling
retry>=0.9.2